Refactor: Implement facade pattern for initial data
